# gerenciamento-estoque/app/Http/Controllers/MovimentacaoRequest.php
<?php

namespace App\Http\Controllers;

/**
 * Class MovimentacaoRequest
 * @package App\Http\Controllers
 */
class MovimentacaoRequest
{
    // This class is used only for Swagger documentation purposes.
}